requirements.txt
streamlit
python-pptx
deepl
openai
python-dotenv
deep-translator
pymupdf
python-docx
